refactor: move gym badge to separate file
