+++ src/components/layout/dialogs/profile/GymBadges.jsx
@@ -1,0 +1,120 @@
+// @ts-check
+import * as React from 'react'
+import { Grid, Typography, IconButton, Button, Box } from '@mui/material'
+import Edit from '@mui/icons-material/Edit'
+import { useTranslation } from 'react-i18next'
+import { useQuery } from '@apollo/client'
+import { useMap } from 'react-leaflet'
+
+import { useLayoutStore, useStatic } from '@hooks/useStore'
+import Query from '@services/Query'
+
+import { VirtualGrid } from '@components/layout/general/VirtualGrid'
+import { Img } from '@components/layout/general/Img'
+
+export function UserGymBadges() {
+  const { t } = useTranslation()
+  /** @type {import('@apollo/client').QueryResult<{ badges: import('@rm/types').Gym[] }>} */
+  const { data } = useQuery(Query.gyms('badges'), {
+    fetchPolicy: 'network-only',
+  })
+
+  const counts = React.useMemo(() => {
+    const counter = { gold: 0, silver: 0, bronze: 0 }
+
+    if (data?.badges) {
+      data.badges.forEach((gym) => {
+        switch (gym.badge) {
+          case 3:
+            counter.gold += 1
+            break
+          case 2:
+            counter.silver += 1
+            break
+          case 1:
+            counter.bronze += 1
+            break
+          default:
+        }
+      })
+    }
+    return counter
+  }, [data])
+
+  return data ? (
+    <Box className="user-profile-badge-grid">
+      <Typography variant="h5" align="center" gutterBottom>
+        {t('gym_badges')}
+      </Typography>
+      <Grid container pt={1} pb={2}>
+        {Object.entries(counts).map(([key, count], i) => (
+          <Grid key={key} item xs={4}>
+            <Typography
+              variant="subtitle2"
+              align="center"
+              className={`badge_${i + 1}`}
+            >
+              {t(`badge_${i + 1}`)}: {count}
+            </Typography>
+          </Grid>
+        ))}
+      </Grid>
+      <VirtualGrid data={data?.badges || []} xs={4} md={3} useWindowScroll>
+        {(_, badge) => <BadgeTile {...badge} />}
+      </VirtualGrid>
+    </Box>
+  ) : null
+}
+
+/** @param {import('@rm/types').Gym} props */
+function BadgeTile({ badge, ...gym }) {
+  const { t } = useTranslation()
+  const map = useMap()
+  const badgeIcon = useStatic((s) => s.Icons.getMisc(`badge_${badge}`))
+
+  return badge ? (
+    <Box className="vgrid-item">
+      <IconButton
+        className="vgrid-icon"
+        disabled={gym.deleted}
+        size="small"
+        onClick={() =>
+          useLayoutStore.setState({
+            gymBadge: { badge, gymId: gym.id, open: true },
+          })
+        }
+      >
+        <Edit />
+      </IconButton>
+      <Button
+        className="vgrid-image"
+        onClick={() => map.flyTo([gym.lat, gym.lon], 16)}
+        disabled={gym.deleted}
+      >
+        <Img
+          className="badge-diamond"
+          src={gym.url ? gym.url.replace('http://', 'https://') : ''}
+          alt={gym.url}
+          height={120}
+          width={120}
+        />
+        {gym.deleted && <div className="disabled-overlay badge-diamond" />}
+        {badge && (
+          <Img
+            src={badgeIcon}
+            alt={t(`badge_${badge}`)}
+            width={96}
+            zIndex={10}
+          />
+        )}
+      </Button>
+      <Typography
+        className="vgrid-caption"
+        variant="caption"
+        color={gym.deleted ? 'GrayText' : 'inherit'}
+      >
+        {gym.name || t('unknown_gym')}
+      </Typography>
+    </Box>
+  ) : null
+}
